tacos on join/leave server moved to join/leave cog
